dom/workers/test/serviceworkers/test_post_message.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 982726 - Test service worker post message </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">
  var magic_value = "MAGIC_VALUE_123";
  function start() {
    return navigator.serviceWorker.register("message_posting_worker.js",
                                            { scope: "./sw_clients/" });
  }

  function testPostMessage(swr) {
    var p = new Promise(function(res, rej) {
      window.onmessage = function(e) {
        if (e.data === "READY") {
          swr.active.postMessage(magic_value);
        } else if (e.data === magic_value) {
          ok(true, "Worker posted the correct value.");
          res();
        } else {
          ok(false, "Wrong value. Expected: " + magic_value +
                    ", got: " + e.data);
          res();
        }
      }
    });

    var w;
    setTimeout(function() {
      w = window.open("sw_clients/service_worker_controlled.html");
    }, 100);
    return p.then(() => w.close());
  }

  function runTest() {
    start()
      .then(testPostMessage).catch(function(e) {
        ok(false, "Some test failed with error " + e);
      }).then(SimpleTest.finish);
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
